Add visualizer for strftime date format
