+++ .gitignore
@@ -1,3 +1,6 @@
 config/
 .vscode/
 __pycache__/
+package*.json
+node_modules/
+!strftime-min.js
