+++ webui/templates/settings.html
@@ -22,6 +22,17 @@
 </div>
 
 {% endblock %} {% block scripts %}
+{% if form.date_format %}
+<script src="{{ url_for('static', filename='node_modules/strftime/strftime-min.js')}}"></script>
+<script>
+  let ex = document.getElementById("example");
+  let format = document.getElementById("date_format");
+  ex.value = strftime(format.value);
+  format.addEventListener("input", (event) => {
+    ex.value = strftime(format.value);
+  });
+</script>
+{% endif %}
 <script>
   try {
     let enabled = document.getElementById("enable_form");
